// src/models/Browser/Runtime.ts
export interface Port {
  name: string;
  disconnect: () => void;
  error: Error;
  onDisconnect: {
    addListener:    (listener: Function) => void;
    removeListener: (listener: Function) => void;
  };
  onMessage: {
    addListener:    (listener: Function) => void;
    removeListener: (listener: Function) => void;
  };
  postMessage: (message: object) => void;

  sender?: any;
}